// front-end/src/pages/PersonalAccount/middleBlock/components/Main/components/types.ts
export interface IAllDeliveryAddress {
  id: string
  deliveryAddress: string
  numberInAddress: string
}
export interface IChangedPassword {
  oldPassword: string
  newPassword: string
  repeatNewPassword: string
}
export interface IChangedInfoData {
  id: string
  name: string
  lastName: string
  email: string
  telefone: string
}